selftests/xsk: Add support check for bpf_xdp_adjust_tail() helper in xskxceiver
Add is_adjust_tail_supported function to check if the
bpf_xdp_adjust_tail() helper is supported when 'adjust_tail' is set in the
test. Look up a specific key in the bss section of the XDP program. If the
key is not found or its value is -EOPNOTSUPP, return false, indicating that
the helper is not supported.

Signed-off-by: Tushar Vyavahare <tushar.vyavahare@intel.com>